requirements.txt
numpy==2.4.2
opencv_python==4.13.0.92
opencv_python_headless==4.10.0.84
Pillow==12.1.1
rfdetr==1.5.0
supervision==0.27.0.post1
torch==2.10.0
ultralytics==8.4.14
